Fixed a stubborn bug and refactoring concering k-deviation pruning

+++ bfs.h
@@ -41,7 +41,7 @@ public:
     std::mutex deviation_write;
     int deviation_pos    = -1; // save more than one?
     int deviation_val    = -1;
-    int deviation_acc    = -1;
+    long deviation_acc    = -1;
     int deviation_vertex = -1;
 
     // memory management
